refactor: split BurstDetector into named phases
Extract the four phases of burst detection into the new
BurstDetectionPhases enum so each phase can be exercised in isolation:

- Timestamps: EXIF date + subsecond extraction
- TimeGrouping: time-proximity bucketing
- Similarity: Vision feature-print verification
- BestSelection: sharpness + aesthetics scoring

BurstDetector keeps its public init, detect(photos:), and
readPreciseTimestamp(filePath:) surface and now composes the phases.
Algorithm, thresholds, comparators, and ordering are unchanged.
